Refactor react app structure and included everything in a single project
