--- react-components/oodt_fm_plugin/webpack.config.js
@@ -1,21 +1,0 @@
-const path = require('path');
-
-module.exports = {
-    entry: './src/index.js',
-    module: {
-        rules: [
-            {
-                test: /\.js$/,
-                use:{
-                    loader: "babel-loader",
-                },
-            },
-        ]
-    },
-    output: {
-        path: path.resolve(__dirname, 'dist/'),
-        publicPath: '',
-        filename: 'build.js',
-        libraryTarget: 'umd'
-    }
-};
