+++ react-components/oodt_opsui_sample_app/package.json
@@ -3,16 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-scripts": "3.0.1",
-    "@material-ui/core": "latest",
+    "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.2.0",
-    "axios": "^0.19.0",
+    "axios": "^0.19.2",
     "clsx": "^1.0.4",
-    "oodt_fm_plugin": "file:../oodt_fm_plugin",
-    "oodt_wm_plugin": "file:../oodt_wm_plugin",
-    "react-router-dom": "^5.0.1"
+    "prop-types": "^15.7.2",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^3.4.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -34,5 +33,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
+  },
+  "description": "ReactJs based OPSUI dashboard for Apache OODT",
+  "main": "index.js",
+  "author": "Apache Software Foundation",
+  "license": "Apache-2.0"
 }
