+++ react-components/oodt_opsui_sample_app/src/components/fileManager/components/Product/index.js
