+++ react-components/oodt_opsui_sample_app/src/components/fileManager/components/ProductList/index.js
