+++ react-components/oodt_opsui_sample_app/src/components/fileManager/constants/fmconnection.js
