+++ react-components/oodt_opsui_sample_app/src/components/workflowManager/constants/wmconnection.js
