--- react-components/oodt_wm_plugin/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "oodt_wm_plugin",
-  "version": "1.0.0",
-  "description": "This is the React Plugin for Apache OODT Workflow Manager",
-  "main": "dist/build.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/apache/oodt/tree/development/react-components/oodt_wm_plugin"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --mode production"
-  },
-  "keywords": [
-    "react",
-    "components",
-    "Apache",
-    "OODT",
-    "Workflow Manager"
-  ],
-  "author": "Nadeeshan Gimhana",
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "@babel/core": "^7.2.2",
-    "@babel/preset-env": "^7.2.3",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
-    "babel-loader": "^8.0.5",
-    "webpack": "^4.28.3",
-    "webpack-cli": "^3.2.0"
-  },
-  "dependencies": {
-    "@material-ui/core": "^3.9.3",
-    "axios": "^0.18.0",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "prop-types": "latest"
-  }
-}
